move test code to test
